Memorymapstorage

+++ src/main/java/io/github/mzmine/util/RawDataFileUtils.java
@@ -1,17 +1,17 @@
 /*
- * Copyright 2006-2018 The MZmine 2 Development Team
+ * Copyright 2006-2020 The MZmine Development Team
  * 
- * This file is part of MZmine 2.
+ * This file is part of MZmine.
  * 
- * MZmine 2 is free software; you can redistribute it and/or modify it under the terms of the GNU
+ * MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
  * General Public License as published by the Free Software Foundation; either version 2 of the
  * License, or (at your option) any later version.
  * 
- * MZmine 2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
+ * MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ * Public License for more details.
  * 
- * You should have received a copy of the GNU General Public License along with MZmine 2; if not,
+ * You should have received a copy of the GNU General Public License along with MZmine; if not,
  * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
  * USA
  */
@@ -31,10 +31,10 @@ import io.github.mzmine.datamodel.Scan;
  */
 public class RawDataFileUtils {
 
-  public static @Nonnull Range<Double> findTotalRTRange(RawDataFile dataFiles[], int msLevel) {
-    Range<Double> rtRange = null;
+  public static @Nonnull Range<Float> findTotalRTRange(RawDataFile dataFiles[], int msLevel) {
+    Range<Float> rtRange = null;
     for (RawDataFile file : dataFiles) {
-      Range<Double> dfRange = file.getDataRTRange(msLevel);
+      Range<Float> dfRange = file.getDataRTRange(msLevel);
       if (dfRange == null)
         continue;
       if (rtRange == null)
@@ -43,7 +43,7 @@ public class RawDataFileUtils {
         rtRange = rtRange.span(dfRange);
     }
     if (rtRange == null)
-      rtRange = Range.singleton(0.0);
+      rtRange = Range.singleton(0.0f);
     return rtRange;
   }
 
@@ -94,5 +94,3 @@ public class RawDataFileUtils {
     return scanNums[best];
   }
 }
-
-
